Calibration Detection Update
Added functions to filter robot image such that the calibration spots are the most prominent feature, and the background is reduced. Added functions to determine whether or not a detected region, post color-mask, is or is not a calibration spot. Added functions to determine position of calibration spots. Added function to determine the scaling unit of pixels to real world units, cm/pix.