Improved robotFilter and length calibration framework
Improved the filterRobotSnapshot by adding a median intensity scaling, which adjust the contrast of the image based on its median intensity. This should help in darker/brighter than expected environments to bring the calibration spots within colormask bounds.

Also updated the length_calibration function with the framework for a calibration that spins the robot to calculate lengths for a variety of angles.

Moved filtration to an external operation on all functions. This means that you'll have to pass the filtered image in to most instead of passing in a raw image and have it filtered in the function itself. Done for efficieny reasons.